feat: re-organize test folders
