+++ API/tests/routes/hello.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-const app = require('../index');
+const app = require('../../index');
 
 describe('GET /api/hello', () => {
     it('should return Hello World', async () => {
